UML-3460 removed unnecessary lpastore person created new files for dedicated tests
