+++ service-api/app/test/AppTest/Service/Lpa/LpaServiceTest.php
@@ -1154,419 +1154,4 @@ class LpaServiceTest extends TestCase
 
         $this->assertEmpty($result);
     }
-
-    #[Test]
-    public function can_serialise_datastore_lpa_to_modernise_format(): void
-    {
-        $lpa = [
-                'uid' => 'M-789Q-P4DF-4UX3',
-                'status' => 'registered',
-                'registrationDate' => '2024-01-12',
-                'updatedAt' => '2024-01-12',
-                'lpaType' => 'personal-welfare',
-                'channel' => 'online',
-                'donor' => [
-                    'uid' => 'eda719db-8880-4dda-8c5d-bb9ea12c236f',
-                    'firstNames' => 'Feeg',
-                    'lastName' => 'Bundlaaaa',
-                    'address' => [
-                        'line1' => '74 Cloob Close',
-                        'town' => 'Mahhhhhhhhhh',
-                        'country' => 'GB',
-                    ],
-                    'email' => 'nobody@not.a.real.domain',
-                    'dateOfBirth' => '1970-01-24',
-                    'contactLanguagePreference' => 'en',
-                    'identityCheck' => [
-                        'checkedAt' => '2024-01-10',
-                        'type' => 'one-login',
-                    ]
-                ],
-                'attorneys'=> [
-                    [
-                        'uid' => '9ac5cb7c-fc75-40c7-8e53-059f36dbbe3d',
-                        'firstNames' => 'Herman',
-                        'lastName' => 'Seakrest',
-                        'address' => [
-                            'line1' => '81 NighOnTimeWeBuiltIt Street',
-                            'town' => 'Mahhhhhhhhhh',
-                            'country' => 'GB',
-                        ],
-                        'status' => 'active',
-                        'channel' => 'paper',
-                        'signedAt' => '2024-01-10',
-                    ],
-                ],
-                'trustCorporations'=> [
-                    [
-                    'uid'=> '1d95993a-ffbb-484c-b2fe-f4cca51801da',
-                    'name'=> 'Trust us Corp.',
-                    'companyNumber'=> '666123321',
-                    'address'=> [
-                        'line1'=> '103 Line 1',
-                        'town'=> 'Town',
-                        'country'=> 'GB',
-                    ],
-                    'status'=> 'active',
-                    'channel'=> 'paper',
-                    'signedAt'=> '2024-01-10',
-                    ],
-                ],
-                'certificateProvider'=> [
-                    'uid' => '6808960d-12cf-47c5-a2bc-3177deb8599c',
-                    'firstNames' => 'Vone',
-                    'lastName' => 'Spust',
-                    'address' => [
-                        'line1'=> '122111 Zonnington Way',
-                        'town' => 'Mahhhhhhhhhh',
-                        'country' => 'GB',
-                    ],
-                    'channel' => 'online',
-                    'email' => 'a@example.com',
-                    'phone' => '070009000',
-                    'signedAt' => '2024-01-10',
-                    'identityCheck' => [
-                        'checkedAt' => '2024-01-10',
-                        'type' => 'one-login',
-                    ],
-                ],
-                'lifeSustainingTreatmentOption'=> 'option-a',
-                'signedAt'=> '2024-01-10',
-                'howAttorneysMakeDecisions'=> 'jointly',
-                'whenTheLpaCanBeUsed'=> 'when-capacity-lost',
-                'applicationHasGuidance'=> false,
-                'applicationHasRestrictions'=> false,
-                'applicationType'=> null,
-            ];
-
-        $expectedLpa =  [
-            "applicationHasGuidance" => false,
-            "applicationHasRestrictions" => false,
-            "applicationType" => null,
-            "attorneyActDecisions" => "jointly",
-            "attorneys" => [
-                [
-                    "name" => null,
-                    "addressLine1" => "81 NighOnTimeWeBuiltIt Street",
-                    "addressLine2" => null,
-                    "addressLine3" => null,
-                    "country" => "GB",
-                    "county" => null,
-                    "postcode" => null,
-                    "town" => "Mahhhhhhhhhh",
-                    "type" => null,
-                    "dob" => null,
-                    "email" => null,
-                    "firstname" => null,
-                    "firstnames" => "Herman",
-                    "surname" => "Seakrest",
-                    "otherNames" => null,
-                    "systemStatus" => "active",
-                ],
-            ],
-            "caseSubtype" => "hw",
-            "channel" => "online",
-            "dispatchDate" => null,
-            "donor" => [
-                "name" => null,
-                "addressLine1" => "74 Cloob Close",
-                "addressLine2" => null,
-                "addressLine3" => null,
-                "country" => "GB",
-                "county" => null,
-                "postcode" => null,
-                "town" => "Mahhhhhhhhhh",
-                "type" => null,
-                "dob" => "1970-01-24T00:00:00+00:00",
-                "email" => "nobody@not.a.real.domain",
-                "firstname" => null,
-                "firstnames" => "Feeg",
-                "surname" => "Bundlaaaa",
-                "otherNames" => null,
-                "systemStatus" => null,
-            ],
-            "hasSeveranceWarning" => null,
-            "invalidDate" => null,
-            "lifeSustainingTreatment" => "option-a",
-            "lpaDonorSignatureDate" => "2024-01-10T00:00:00+00:00",
-            "lpaIsCleansed" => null,
-            "onlineLpaId" => null,
-            "receiptDate" => null,
-            "registrationDate" => "2024-01-12T00:00:00+00:00",
-            "rejectedDate" => null,
-            "replacementAttorneys" => null,
-            "status" => "registered",
-            "statusDate" => null,
-            "trustCorporations" => [
-                [
-                    "name" => "Trust us Corp.",
-                    "addressLine1" => "103 Line 1",
-                    "addressLine2" => null,
-                    "addressLine3" => null,
-                    "country" => "GB",
-                    "county" => null,
-                    "postcode" => null,
-                    "town" => "Town",
-                    "type" => null,
-                    "dob" => null,
-                    "email" => null,
-                    "firstname" => null,
-                    "firstnames" => null,
-                    "surname" => null,
-                    "otherNames" => null,
-                    "systemStatus" => "active",
-                    "companyName" => "Trust us Corp.",
-                ],
-            ],
-            "uId" => null,
-            "withdrawnDate" => null,
-        ];
-
-        $newLpa = ($this->lpaDataFormatter)($lpa);
-
-        $jsonLpa = json_encode($newLpa);
-        $expectedJsonLpa = json_encode($expectedLpa);
-    print_r($jsonLpa);
-        $this->assertEquals($expectedJsonLpa, $jsonLpa);
-    }
-
-    #[Test]
-    public function can_cast_single_donor(): void
-    {
-        $donor = [
-            'uId' => '700000000971',
-            'firstNames' => 'Feeg',
-            'surname' => 'Gilson',
-            'lastName' => 'Bundlaaaa',
-            'email' => 'nobody@not.a.real.domain',
-            'middlenames' => 'Suzanne',
-            'dateOfBirth' => '1970-01-24',
-            'address' => [
-                'line1' => '74 Cloob Close',
-                'town' => 'Mahhhhhhhhhh',
-                'country' => 'GB',
-            ],
-        ];
-
-        $expectedDatastoreDonor = new LpaStoreDonor(
-            null,
-            '74 Cloob Close',
-            null,
-            null,
-            'GB',
-            null,
-            null,
-            'Mahhhhhhhhhh',
-            null,
-            new \DateTimeImmutable('1970-01-24 00:00:00.000000'),
-            'nobody@not.a.real.domain',
-            null,
-            'Feeg',
-            'Bundlaaaa',
-            null,
-            null,
-        );
-
-        $castSingleDonor = new CastSingleDonor();
-
-        $mapper = new ObjectMapperUsingReflection(
-            new DefinitionProvider(
-                keyFormatter: new KeyFormatterWithoutConversion(),
-            ),
-        );
-
-        $result = $castSingleDonor->cast($donor, $mapper);
-        $this->assertEquals($expectedDatastoreDonor, $result);
-    }
-
-    #[Test]
-    public function can_cast_case_subtype(): void
-    {
-        $caseSubType = 'personal-welfare';
-
-        $expectedCaseSubType = 'hw';
-
-        $castToSingleSubType = new CastToCaseSubtype();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $castToSingleSubType->cast($caseSubType, $mockHydrator);
-
-        $this->assertEquals($expectedCaseSubType, $result);
-    }
-
-    #[Test]
-    public function can_cast_life_sustaining_treatment(): void
-    {
-        $lifeSustainingTreatment = 'option-a';
-
-        $expectedLifeSustainingTreatment = 'option-a';
-
-        $castToLifeSustainingTreatment = new CastToLifeSustainingTreatment();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $castToLifeSustainingTreatment->cast($lifeSustainingTreatment, $mockHydrator);
-
-        $this->assertEquals($expectedLifeSustainingTreatment, $result);
-    }
-
-    #[Test]
-    public function can_when_lpa_can_be_used(): void
-    {
-        $whenTheLpaCanBeUsed = 'singular';
-
-        $expectedWhenTheLpaCanBeUsed = 'singular';
-
-        $castToWhenTheLpaCanBeUsed = new CastToWhenTheLpaCanBeUsed();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $castToWhenTheLpaCanBeUsed->cast($whenTheLpaCanBeUsed, $mockHydrator);
-
-        $this->assertEquals($expectedWhenTheLpaCanBeUsed, $result);
-    }
-
-    #[Test]
-    public function can_date_to_string_serialised(): void
-    {
-        $date = new \DateTimeImmutable('22-12-1997');
-
-        $expecteDate = '1997-12-22T00:00:00+00:00';
-
-        $castDateToStringSerialize = new DateToStringSerializer();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $castDateToStringSerialize->serialize($date, $mockHydrator);
-
-        $this->assertEquals($expecteDate, $result);
-    }
-
-    #[Test]
-    public function cannot_convert_date_to_string(): void
-    {
-        $date = '22-12-1997';
-
-        $expecteDate = '22-12-1997';
-
-        $castDateToStringSerialize = new DateToStringSerializer();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $castDateToStringSerialize->serialize($date, $mockHydrator);
-
-        $this->assertEquals($expecteDate, $result);
-    }
-
-    #[Test]
-    public function can_extract_town_from_datastore(): void
-    {
-        $address = [
-            'line1'   => '74 Cloob Close',
-            'town'    => 'Mahhhhhhhhhh',
-            'country' => 'GB',
-        ];
-
-        $expectedTown = 'Mahhhhhhhhhh';
-
-        $extractTownFromDataStore = new ExtractTownFromLpaStore();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $extractTownFromDataStore->cast($address, $mockHydrator);
-
-        $this->assertEquals($expectedTown, $result);
-    }
-
-    #[Test]
-    public function cannot_extract_town_from_datastore(): void
-    {
-        $address = [
-            'line1'   => '74 Cloob Close',
-            'country' => 'GB',
-        ];
-
-        $extractTownFromDataStore = new ExtractTownFromLpaStore();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $extractTownFromDataStore->cast($address, $mockHydrator);
-
-        $this->assertEquals(null, $result);
-    }
-
-    #[Test]
-    public function can_extract_country_from_datastore(): void
-    {
-        $address = [
-            'line1'   => '74 Cloob Close',
-            'town'    => 'Mahhhhhhhhhh',
-            'country' => 'GB',
-        ];
-
-        $expectedCountry = 'GB';
-
-        $extractCountryFromDataStore = new ExtractCountryFromLpaStore();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $extractCountryFromDataStore->cast($address, $mockHydrator);
-
-        $this->assertEquals($expectedCountry, $result);
-    }
-
-    #[Test]
-    public function cannot_extract_country_from_datastore(): void
-    {
-        $address = [
-            'line1'   => '74 Cloob Close',
-            'town'    => 'Mahhhhhhhhhh',
-        ];
-
-        $extractCountryFromDataStore = new ExtractCountryFromLpaStore();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $extractCountryFromDataStore->cast($address, $mockHydrator);
-
-        $this->assertEquals(null, $result);
-    }
-
-    #[Test]
-    public function can_extract_address_one_from_datastore(): void
-    {
-        $address = [
-            'line1'   => '74 Cloob Close',
-            'town'    => 'Mahhhhhhhhhh',
-            'country' => 'GB',
-        ];
-
-        $expectedAddressOne = '74 Cloob Close';
-
-        $extractAddressOneFromDataStore = new ExtractAddressLine1FromLpaStore();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $extractAddressOneFromDataStore->cast($address, $mockHydrator);
-
-        $this->assertEquals($expectedAddressOne, $result);
-    }
-
-    #[Test]
-    public function cannot_extract_address_one_from_datastore(): void
-    {
-        $address = [
-            'town'    => 'Mahhhhhhhhhh',
-            'country' => 'GB',
-        ];
-
-        $extractAddressOneFromDataStore = new ExtractAddressLine1FromLpaStore();
-
-        $mockHydrator = $this->createMock(ObjectMapper::class);
-
-        $result = $extractAddressOneFromDataStore->cast($address, $mockHydrator);
-
-        $this->assertEquals(null, $result);
-    }
 }
